requirements.txt
apache-airflow
pandas
pyarrow
boto3
s3fs
requests
